Fixed buy track, and tests
